+++ helper/db.js
@@ -22,13 +22,23 @@ module.exports = (callback) => {
 
             // table init
             const createTableQuery = "CREATE TABLE IF NOT EXISTS files (" +
-                "id INT PRIMARY KEY NOT NULL," +
+                "id INT PRIMARY KEY NOT NULL AUTO_INCREMENT," +
                 "value TEXT);";
             con.query(createTableQuery, (err, res) => {
-                //TODO handle errors
+                // TODO handle errors
             });
+
+
+            // --- TODO (TO REMOVE) single file management
+            const tempTruncateQuery = "TRUNCATE TABLE files;";
+            con.query(tempTruncateQuery, (err, res) => {
+            });
+            const tempInsertQuery = "INSERT INTO files (value) VALUES (\"\");";
+            con.query(tempInsertQuery, (err, res) => {
+            });
+            // ---
 
             callback(con);
         }
     });
-}
+};
